% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sym_kl_distance}
\alias{sym_kl_distance}
\title{Symmetric KL divergence between diagonal Gaussians}
\usage{
sym_kl_distance(mu, var, mu2, var2)
}
\arguments{
\item{mu}{n1 x d matrix of means}

\item{var}{n1 x d matrix of variances}

\item{mu2}{n2 x d matrix of means}

\item{var2}{n2 x d matrix of variances}
}
\value{
n1 x n2 matrix of symmetric KL divergences
}
\description{
sym_KL(i,j) = 0.5 * ( sum_d var[i,d]/var2[j,d] + sum_d var2[j,d]/var[i,d]
                      + mahal_ij[i,j] + mahal_ji[i,j] )
}
\details{
mahal_ij[i,j] = sum_d (mu[i,d]-mu2[j,d])^2 / var2[j,d]
}
