% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native-wrappers.R
\name{w2_distance}
\alias{w2_distance}
\title{Diagonal 2-Wasserstein distance (fully vectorized via BLAS)}
\usage{
w2_distance(mu, var, mu2, var2)
}
\arguments{
\item{mu}{n1 x d matrix of means}

\item{var}{n1 x d matrix of variances (must be > 0)}

\item{mu2}{n2 x d matrix of means}

\item{var2}{n2 x d matrix of variances (must be > 0)}
}
\value{
n1 x n2 matrix of squared distances
}
\description{
Diagonal 2-Wasserstein distance (fully vectorized via BLAS)
}
\keyword{internal}
