% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{geneSetEmbedding-package}
\alias{geneSetEmbedding}
\alias{geneSetEmbedding-package}
\title{geneSetEmbedding package}
\description{
Compute network diffusion features, learn gene embeddings, derive gene set Gaussian representations, and support similarity and concise gene set construction.
}
\author{
\strong{Maintainer}: enrichit contributors \email{noreply@example.com}

}
\keyword{internal}
