% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{gsemb_get_concise_gene_sets}
\alias{gsemb_get_concise_gene_sets}
\title{Create concise gene sets (high-level wrapper)}
\usage{
gsemb_get_concise_gene_sets(x, gene_sets, ...)
}
\arguments{
\item{x}{A `gsemb_embedding` object.}

\item{gene_sets}{Named list of gene sets.}

\item{...}{Passed to `gsemb_concise_gene_sets`.}
}
\value{
A named list of concise gene sets.
}
\description{
Create concise gene sets (high-level wrapper)
}
