% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native-wrappers.R
\name{sym_kl_distance}
\alias{sym_kl_distance}
\title{Symmetric KL divergence between diagonal Gaussians}
\usage{
sym_kl_distance(mu, var, mu2, var2)
}
\arguments{
\item{mu}{n1 x d matrix of means}

\item{var}{n1 x d matrix of variances}

\item{mu2}{n2 x d matrix of means}

\item{var2}{n2 x d matrix of variances}
}
\value{
n1 x n2 matrix of symmetric KL divergences
}
\description{
Symmetric KL divergence between diagonal Gaussians
}
\keyword{internal}
